Refactoring to add more versatility for future developments.
Message scheduling is not hard coded anymore, but it makes use of the class UserAlert, which represents a alert to be scheduled. The installNotifier method now loops through all the UserAlerts of a chatId and installs a notifier for each of them.
At the moment the only hard-coded method is getUserAlerts (it returns a alert for sunrise and sunset), which will retrieve all UserAlerts dynamically from the database, in the future.